feat: common outlet added
