+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
+import { BrowserRouter as Router, Link } from 'react-router-dom';
 import './App.css';
 
-import AppRoutes from './routes/userRoutes';
+import AppRoutes from './routes/appRoutes';
 
 // Adjusted import path if needed
 import './index.css';
@@ -10,17 +10,8 @@ import './index.css';
 const App: React.FC = () => {
   return (
     <Router>
-      <Link to="/"></Link>
-      <Link to="/registerEmail"></Link>
-      <Link to="/verifyOtp"></Link>
-      <Link to="/registerProfile"></Link>
-      <Link to="/registerCompanyProfile"></Link>
-      <Link to="/choosePlan"></Link>
-
-      <div>
         <AppRoutes />
-      </div>
-    </Router>
+     </Router>
   );
 };
